requirementtext.txt
fastapi
sqlalchemy
pymysql
load_dotenv
uvicorn



